🌟 Implement message deletion handling in channel management with logging
